refactor: move to hook
